Moved structs to new file
