+++ camera_manager/include/camera_manager/camera_manager.hpp
@@ -30,39 +30,11 @@ using CameraInfoMsg = sensor_msgs::msg::CameraInfo;
 using StereoSyncPolicy = message_filters::sync_policies::ApproximateTime<ImageMsg, CameraInfoMsg, ImageMsg, CameraInfoMsg>;
 using MonoSyncPolicy = message_filters::sync_policies::ApproximateTime<ImageMsg, CameraInfoMsg>;
 
-struct CameraIntrinsics {
-  cv::Mat K; // Camera intrinsic matrix: focal length, principal point..
-  cv::Mat D; // Camera distortion coefficients
-};
-
-struct Frame {
-  int frameID;
-  rclcpp::Time stamp;
-  cv::Mat image;
-  std::vector<cv::KeyPoint> keypoints;
-  CameraIntrinsics intrinsics; // Camera intrinsics
-  cv::Mat descriptors; // SIFT/ORB descriptors
-};
-
 struct Edge {
   int fromID;
   int toID;
   cv::Mat relativePose; // 4x4 SE(3) Transformation matrix T_from_to
   cv::Mat covariance; // Covariance of the relative pose
-};
-
-struct Feature {
-  int frameID;
-  std::vector<cv::KeyPoint> keypoints;
-  cv::Mat descriptors; // SIFT/ORB descriptors
-};
-
-struct StereoFeature {
-  int frameID;
-  std::vector<cv::KeyPoint> leftKeypoints;
-  std::vector<cv::KeyPoint> rightKeypoints;
-  cv::Mat leftDescriptors; // SIFT/ORB descriptors for left image
-  cv::Mat rightDescriptors; // SIFT/ORB descriptors for right image
 };
 
 class CameraManager : public rclcpp::Node {
